reformat and RSSI

+++ src/driver/zc1103/inc/rfsystem.h
@@ -367,7 +367,7 @@ public:
   void scanR();
 
   /// when measuring RSSI, the chip will enter RX mode
-  uint8_t rssi();
+  int rssi();
 
   uint8_t pollTxPktSt();
